Setting fileformat to unix...
This displays carriage-returns as '^R' markers, which allows you to more
readily spot the fact that the file format is not unix.

An alternative to this would be to specify fileformats=unix,win,
and then use something like Marty Grenfel's status-line to display a
warning if the file isn't in unix format... but I don't really see the
necessity for it.